+++ src/main/java/io/fabric8/maven/docker/config/BuildImageConfiguration.java
@@ -168,6 +168,9 @@ public class BuildImageConfiguration implements Serializable {
     private Boolean skip;
 
     @Parameter
+    private Boolean skipPush;
+
+    @Parameter
     private ArchiveCompression compression = ArchiveCompression.none;
 
     @Parameter
@@ -335,6 +338,10 @@ public class BuildImageConfiguration implements Serializable {
         return skip != null ? skip : false;
     }
 
+    public boolean skipPush() {
+        return skipPush != null ? skipPush : false;
+    }
+
     public Boolean getNoCache() {
         return noCache != null ? noCache : nocache;
     }
@@ -349,6 +356,10 @@ public class BuildImageConfiguration implements Serializable {
 
     public Boolean getSkip() {
         return skip;
+    }
+
+    public Boolean getSkipPush() {
+        return skipPush;
     }
 
     public ArchiveCompression getCompression() {
@@ -599,6 +610,11 @@ public class BuildImageConfiguration implements Serializable {
 
         public Builder skip(Boolean skip) {
             config.skip = skip;
+            return this;
+        }
+
+        public Builder skipPush(Boolean skipPush) {
+            config.skipPush = skipPush;
             return this;
         }